fix edge case where queue processing gets confused by fast blocks
If many blocks get processed quickly (which only tends to happen in
testing), sometimes the queue manager doesn't process queued items
correctly. This fixes the issues by adding a check to make sure the
queue is reprocessed if there are mined transactions in a block
greater than the one currently being processed.